requirements.txt
anyio==3.7.1
bleach==6.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
click-log==0.4.0
colorama==0.4.6
dataclasses==0.6
deprecation==2.1.0
docutils==0.20.1
dotty-dict==1.3.1
exceptiongroup==1.3.0
gitdb==4.0.12
GitPython==3.1.45
gotrue==0.5.4
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
invoke==1.7.3
jaraco.classes==3.2.3
keyring==24.1.1
more-itertools==9.1.0
packaging==24.0
pkginfo==1.10.0
postgrest-py==0.10.4
pydantic==1.10.22
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-gitlab==3.15.0
python-semantic-release==7.32.1
readme-renderer==37.3
realtime==0.0.5
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
semver==2.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
storage3==0.3.6
StrEnum==0.4.15
supabase==0.7.1
supafunc==0.2.2
tomlkit==0.12.5
tqdm==4.67.1
twine==3.8.0
typing_extensions==4.7.1
urllib3==2.0.7
webencodings==0.5.1
websockets==10.4
zipp==3.15.0
